% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiled_code_tslstmx.R
\name{initialize_tensorflow}
\alias{initialize_tensorflow}
\title{Function to initialize TensorFlow and enable eager execution}
\usage{
initialize_tensorflow()
}
\value{
No return value, called for smooth running
}
\description{
This function initializes TensorFlow and enables eager execution.
}
\examples{
\donttest{
initialize_tensorflow()
}

}
